requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching and search
redis==5.0.1
elasticsearch==8.11.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and ML
sentence-transformers==2.2.2
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
chromadb==0.4.18
spacy==3.7.2

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Configuration and utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Synthetic data generation
faker==21.0.0
names==0.3.0
